# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Beacon": {
        "additionalProperties": false,
        "properties": {
          "duration_ms": {
            "anyOf": [
              {
                "maximum": 86400000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          }
        },
        "title": "Beacon",
        "type": "object"
      },
...
      "EntityRef": {
        "additionalProperties": false,
        "properties": {
          "entity_id": {
            "format": "uuid",
            "title": "Entity Id",
            "type": "string"
          },
          "entity_type": {
            "pattern": "^[a-z][a-z_]{1,39}$",
            "title": "Entity Type",
            "type": "string"
          }
        },
        "required": [
          "entity_type",
          "entity_id"
        ],
        "title": "EntityRef",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LinkCreate": {
        "additionalProperties": false,
        "properties": {
          "entity_id": {
            "format": "uuid",
            "title": "Entity Id",
            "type": "string"
          },
          "entity_type": {
            "pattern": "^[a-z][a-z_]{1,39}$",
            "title": "Entity Type",
            "type": "string"
          },
          "expires_in_days": {
            "anyOf": [
              {
                "maximum": 365.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires In Days"
          },
          "scopes": {
            "items": {
              "enum": [
                "view",
                "accept",
                "pay"
              ],
              "type": "string"
            },
            "minItems": 1,
            "title": "Scopes",
            "type": "array"
          },
          "send_to": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SendTo"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "entity_type",
          "entity_id"
        ],
        "title": "LinkCreate",
        "type": "object"
      },
      "LinkCreated": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "entity_id": {
            "format": "uuid",
            "title": "Entity Id",
            "type": "string"
          },
          "entity_type": {
            "title": "Entity Type",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_viewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Viewed At"
          },
          "message_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Outbound email, if `send_to` was given",
            "title": "Message Id"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "title": "Scopes",
            "type": "array"
          },
          "token": {
            "title": "Token",
            "type": "string"
          },
          "url": {
            "description": "Share this. The token is shown only once and stored only as a hash.",
            "title": "Url",
            "type": "string"
          },
          "view_count": {
            "title": "View Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "entity_type",
          "entity_id",
          "scopes",
          "expires_at",
          "revoked_at",
          "created_at",
          "view_count",
          "last_viewed_at",
          "url",
          "token"
        ],
        "title": "LinkCreated",
        "type": "object"
      },
      "LinkEventOut": {
        "properties": {
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_bot": {
            "title": "Is Bot",
            "type": "boolean"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "event_type",
          "occurred_at",
          "is_bot",
          "details"
        ],
        "title": "LinkEventOut",
        "type": "object"
      },
...
      "PublicLinkView": {
        "description": "What an anonymous visitor may learn about a link. No internal ids.",
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "has_download": {
            "title": "Has Download",
            "type": "boolean"
          },
          "has_web_view": {
            "title": "Has Web View",
            "type": "boolean"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "title": "Scopes",
            "type": "array"
          },
          "tenant": {
            "$ref": "#/components/schemas/PublicTenant"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "tenant",
          "title",
          "kind",
          "scopes",
          "expires_at",
          "has_web_view",
          "has_download"
        ],
        "title": "PublicLinkView",
        "type": "object"
      },
      "PublicTenant": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "PublicTenant",
        "type": "object"
      },
...
      "SendTo": {
        "additionalProperties": false,
        "description": "Email the link to the client right away (logged in the outbound messages).",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "message": {
            "default": "",
            "maxLength": 2000,
            "title": "Message",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "email",
          "name"
        ],
        "title": "SendTo",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "app__modules__documents__schemas__LinkOut": {
        "properties": {
          "entity_id": {
            "format": "uuid",
            "title": "Entity Id",
            "type": "string"
          },
          "entity_type": {
            "title": "Entity Type",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "entity_type",
          "entity_id"
        ],
        "title": "LinkOut",
        "type": "object"
      },
...
      "app__modules__public_links__schemas__LinkOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "entity_id": {
            "format": "uuid",
            "title": "Entity Id",
            "type": "string"
          },
          "entity_type": {
            "title": "Entity Type",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_viewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Viewed At"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "title": "Scopes",
            "type": "array"
          },
          "view_count": {
            "title": "View Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "entity_type",
          "entity_id",
          "scopes",
          "expires_at",
          "revoked_at",
          "created_at",
          "view_count",
          "last_viewed_at"
        ],
        "title": "LinkOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "BrokerOS API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/documents/{document_id}/links": {
      "post": {
        "operationId": "documents_link",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityRef"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__modules__documents__schemas__LinkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Link",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{document_id}/links/{link_id}": {
      "delete": {
        "operationId": "documents_unlink",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "link_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Link Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Link",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/api/v1/public-links": {
      "get": {
        "operationId": "public_links_list",
        "parameters": [
          {
            "in": "query",
            "name": "entity_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 40,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Type"
            }
          },
          {
            "in": "query",
            "name": "entity_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/app__modules__public_links__schemas__LinkOut"
                  },
                  "title": "Response Public Links List",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Links",
        "tags": [
          "public links"
        ]
      },
      "post": {
        "operationId": "public_links_create",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Link",
        "tags": [
          "public links"
        ]
      }
    },
    "/api/v1/public-links/{link_id}/events": {
      "get": {
        "operationId": "public_links_events",
        "parameters": [
          {
            "in": "path",
            "name": "link_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Link Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LinkEventOut"
                  },
                  "title": "Response Public Links Events",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Link Events",
        "tags": [
          "public links"
        ]
      }
    },
    "/api/v1/public-links/{link_id}/revoke": {
      "post": {
        "description": "Takes effect immediately: the next request with the token gets 410.",
        "operationId": "public_links_revoke",
        "parameters": [
          {
            "in": "path",
            "name": "link_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Link Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__modules__public_links__schemas__LinkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke",
        "tags": [
          "public links"
        ]
      }
    },
    "/api/v1/public/links/{token}": {
      "get": {
        "operationId": "public_link_get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 10,
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicLinkView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Public",
        "tags": [
          "public"
        ]
      }
    },
    "/api/v1/public/links/{token}/beacon": {
      "post": {
        "description": "Sent by the public page after it rendered. Only this counts as a view; bots are recorded, not counted.",
        "operationId": "public_link_beacon",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 10,
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Beacon"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Beacon",
        "tags": [
          "public"
        ]
      }
    },
    "/api/v1/public/links/{token}/download": {
      "get": {
        "operationId": "public_link_download",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 10,
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download",
        "tags": [
          "public"
        ]
      }
    },
    "/api/v1/public/links/{token}/html": {
      "get": {
        "description": "Self-contained HTML for a sandboxed iframe (no scripts, no network; strict CSP).",
        "operationId": "public_link_html",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 10,
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Public Html",
        "tags": [
          "public"
        ]
      }
    },
...
  }
}